# fiber-backend/docs/swagger.yaml
basePath: /api
definitions:
  influx.Point:
    properties:
      chamber_id:
        type: string
      destination:
        type: string
      field:
        type: string
      layer_id:
        type: string
      measurement:
        type: string
      source:
        type: string
      system:
        type: string
      telegraf_instance_id:
        type: string
      time:
        type: string
      value: {}
      wafer_id:
        type: string
    type: object
  user.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        description: access token lifetime in seconds
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  user.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  user.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        minLength: 2
        type: string
      role:
        type: string
      updated_at:
        type: string
    required:
    - email
    - name
    type: object
  user.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Go Fiber v3 backend with JWT authentication, Argon2 password hashing,
    and PostgreSQL.
  title: Fiber Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns access + refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Log in with email and password
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the provided refresh token, preventing further use
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Log out (invalidate refresh token)
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's profile data (requires JWT)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access + refresh token
        pair (token rotation)
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with Argon2id hashed password and returns
        access + refresh tokens
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: Validation error or bad request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Email already registered
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /influx/health:
    get:
      consumes:
      - application/json
      description: Check if InfluxDB is reachable
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: InfluxDB Health
      tags:
      - influx
  /influx/range:
    get:
      consumes:
      - application/json
      description: Query points from InfluxDB for a specific measurement with optional
        tag filters (last 1h)
      parameters:
      - description: Measurement name
        in: query
        name: measurement
        required: true
        type: string
      - default: '"-1h"'
        description: Range (e.g., -1h, -5m)
        in: query
        name: range
        type: string
      - default: 100
        description: Limit number of points
        in: query
        name: limit
        type: integer
      - description: Chamber ID
        in: query
        name: chamber_id
        type: string
      - description: Layer ID
        in: query
        name: layer_id
        type: string
      - description: Wafer ID
        in: query
        name: wafer_id
        type: string
      - description: System
        in: query
        name: system
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/influx.Point'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Query InfluxDB range
      tags:
      - influx
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.UserResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Admin endpoint to create a user with specified fields
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user (admin)
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token. The `Bearer ` prefix is optional.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"